cli: share resource web api wrapper
